pyproject.toml
[project]
name = "qwen3-pytorch"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.4",
    "safetensors>=0.6.2",
    "tokenizers>=0.22.1",
]
